Move `_bounds_set_from_proto` to the bounds proto module
`_bounds_set_from_proto` converts a sequence of protobuf `Bounds` into
a `BoundsSet | InvalidBoundsSet`. That is a bounds converter, so its
natural home is `_bounds.py` next to `bounds_from_proto2` (which it
calls), not `_sample.py`, where it only happened to be first needed.

Relocating it lets other converters reuse it without reaching into the
sample module — in particular the electrical-component metric config
bounds converter, which is about to aggregate duplicate metrics through
the same helper.
